코드 정리

+++ Assets/Scripts/Enemy/Boss/Queen.cs
@@ -11,7 +11,7 @@ public class Queen : MonoBehaviour
 
     public void NightTrip() //"종속자를 2명 소환한다 <sprite=11>를 5턴 동안 2 얻는다."
     {
-        if(queenHealth == null)
+        if (queenHealth == null)
         {
             queenHealth = GetComponent<EnemyHealth>();
         }
@@ -24,8 +24,8 @@ public class Queen : MonoBehaviour
 
         skillEvent = (sp, action) =>
         {
-            print((float)queenHealth.maxHp / (float)queenHealth.curHp);
-            if ((float)queenHealth.maxHp / (float)queenHealth.curHp >= 2.0f)
+            print(queenHealth.maxHp / (float)queenHealth.curHp);
+            if (queenHealth.maxHp / (float)queenHealth.curHp >= 2.0f)
             {
                 bh.battleEvent.RemoveEventInfo(skillEventInfo);
